Project/translator-backend/pyproject.toml
[tool.poetry]
name = "translator-backend"
version = "0.1.0"
description = ""
authors = ["Madiyar Mukushev <madiyar2000@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
requests = "^2.27.1"
google-cloud-translate = "^3.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
